tools/tool-base/src/test/java/io/spine/code/gen/js/TypeNameTest.java
/*
 * Copyright 2020, TeamDev. All rights reserved.
 *
 * Redistribution and use in source and/or binary forms, with or without
 * modification, must retain the above copyright notice and the following
 * disclaimer.
 *
 * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS
 * "AS IS" AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT
 * LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR
 * A PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT
 * OWNER OR CONTRIBUTORS BE LIABLE FOR ANY DIRECT, INDIRECT, INCIDENTAL,
 * SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT
 * LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
 * DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY
 * THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
 * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 * OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */

package io.spine.code.gen.js;

import com.google.common.testing.NullPointerTester;
import com.google.protobuf.Any;
import com.google.protobuf.Descriptors.Descriptor;
import com.google.protobuf.Descriptors.EnumDescriptor;
import com.google.protobuf.Syntax;
import org.junit.jupiter.api.DisplayName;
import org.junit.jupiter.api.Test;

import static io.spine.testing.DisplayNames.NOT_ACCEPT_NULLS;
import static org.junit.jupiter.api.Assertions.assertEquals;

@DisplayName("TypeName should")
class TypeNameTest {

    @Test
    @DisplayName(NOT_ACCEPT_NULLS)
    void passNullToleranceCheck() {
        new NullPointerTester().testAllPublicStaticMethods(TypeName.class);
    }

    @Test
    @DisplayName("append `proto.` prefix to message type")
    void appendPrefixToMessage() {
        Descriptor descriptor = Any.getDescriptor();
        TypeName typeName = TypeName.from(descriptor);
        String expected = "proto.google.protobuf.Any";
        assertEquals(expected, typeName.value());
    }

    @Test
    @DisplayName("append `proto.` prefix to enum type")
    void appendPrefixToEnum() {
        EnumDescriptor descriptor = Syntax.getDescriptor();
        TypeName typeName = TypeName.from(descriptor);
        String expected = "proto.google.protobuf.Syntax";
        assertEquals(expected, typeName.value());
    }

    @Test
    @DisplayName("provide a name for a message parser")
    void messageParserName() {
        TypeName anyParser = TypeName.ofParser(Any.getDescriptor());
        assertEquals("proto.google.protobuf.Any.Parser", anyParser.value());
    }
}
